Add the essential queue_request() methods
Adds the queue_request() method to sequences, requests, and gaps.
These methods work together to ensure that all requests in a
sequence are added to an appropriate queue (which must implement
the put() method) in the correct order and with the correct time
gaps.  This completes the work on request.py.