Allow the driver processes to exit on their own
Sometimes the SIGTERM was coming in before the driver was finished
processing its last request, so no response was logged.  The
solution to this is two-fold: first, a soft stop signaling
mechanism is added (place 'STOP' on the queue); second, we add one
last time.sleep(1) prior to killing all the servers.

This also puts exception handling around the os.kill() call, so if
a process has already exited, we'll keep going, thus making sure
all of the servers get a SIGTERM.